ensure data files on startup

+++ lib.go
@@ -1,13 +1,13 @@
 package main
 
 import (
-	"strings"
 	"errors"
 	"fmt"
 	"io/ioutil"
 	"math"
 	"math/rand"
 	"os"
+	"strings"
 	"time"
 )
 
@@ -15,12 +15,24 @@ import (
 func CreateFileIfNotExists(fname string) error {
 	var _, err = os.Stat(fname)
 	// create log file if not exists
-	if os.IsNotExist(err) {
+	if err != nil && os.IsNotExist(err) {
 		file, err := os.Create(fname)
 		if err != nil {
 			return err
 		}
 		defer file.Close()
+	}
+	return nil
+}
+
+// CreateDirIfNotExists creates directory path if not exists
+func CreateDirIfNotExists(dpath string) error {
+	var _, err = os.Stat(dpath)
+	if err != nil && os.IsNotExist(err) {
+		err := os.MkdirAll(dpath, os.ModePerm)
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
